requirements.txt
opencv_python==4.9.0.80
pillow==10.3.0
rembg==2.0.57
matplotlib==3.9.0
numpy==1.26.4
tqdm==4.66.4
torch==2.3.0
torchvision==0.18.0
dl_ext==1.3.4
easydict==1.13

segment_anything @ git+https://github.com/facebookresearch/segment-anything.git
